DAO-181 Rename pages/claims -> pages/my-claims
