+++ src/pages/my-claims/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './my-claims';
